refactor(admin): full-width board on the Rundown, order desk beneath it
Drop the On air listing section entirely — the board already shows the whole
week, and the listing doubled it at the cost of half the width. The right
rail becomes the order desk: a full-width band under the board holding the
sentence editor, the orders behind the edited day, the week-in-numbers,
suggestions, and the airtime bars in four columns. Board day columns now
stretch (min 188px, flex) so seven days fill the page; picking a block on
the board scrolls the desk into view when it is below the fold.
